fix any type convert to api-inerface
